[Fixed] Fixed the visibility check, Prevents enemies from becoming visible when teammates see them.

+++ CS2_External/Entity.cpp
@@ -97,7 +97,7 @@ bool PlayerPawn::GetCameraPos()
 
 bool PlayerPawn::GetSpotted()
 {
-	return GetDataAddressWithOffset<int>(Address, Offset::Pawn.bSpottedByMask, this->bSpottedByMask);
+	return GetDataAddressWithOffset<DWORD64>(Address, Offset::Pawn.bSpottedByMask, this->bSpottedByMask);
 }
 
 bool PlayerPawn::GetWeaponName()
