+++ CS2_External/Entity.h
@@ -49,7 +49,7 @@ public:
 	int Health;
 	int TeamID;
 	int Fov;
-	int bSpottedByMask;
+	DWORD64 bSpottedByMask;
 	int fFlags;
 public:
 	bool GetPos();
